+++ stm32/audio.h
@@ -1,0 +1,5 @@
+#ifndef AUDIO_H
+#define AUDIO_H
+void audio_init_converters();
+void audio_init_bus();
+#endif
